pyproject.toml
[project]
name = "ttc-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.44",
    "gtfs-realtime-bindings>=2.0.0",
    "polars>=1.37.1",
    "pre-commit>=4.5.1",
    "pyzmq>=27.1.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "apache-airflow-task-sdk>=1.1.7",
    "apache-airflow[docker]>=3.1.7",
    "marimo>=0.19.8",
    "ruff>=0.15.0",
    "ty>=0.0.15",
]
